Added NPC teleportation capability
Changed the TeleportImmediatelyTo() function to call or TeleportPlayer() (previous default behaviour) or TeleportNPC(). TeleportPlayer is a extracted method from TeleportImmediatelyTo().

Changed the Start method to make teleporting the object at the start of the monobehaviour to be depenadnt on the TeleportOnStart boolean field which can be edited in the inspector.